REFACT:Update logOut feature(#69)
유저 타입을 구분하여 알맞는 로그아웃 api 호출
